backend/requirements.txt
fastapi
uvicorn
pydub
pyyaml
python-multipart